API: moving into a more traditional structure, doxigen config file update
