+++ Test/calibrate/calibrateApp2.c
@@ -1,11 +1,10 @@
-
 #include <stdio.h>
 #include <stdlib.h>
 #include <termios.h>
 #include <unistd.h>
 
 #include "redpitaya/rp.h"
-#include "../../api/rpbase/src/common.h"
+#include "../../api/src/common.h"
 
 
 void waitForUser ( void )
