+++ api/Doxyfile
@@ -7,8 +7,7 @@ DOXYFILE_ENCODING      = UTF-8
 PROJECT_NAME           = "Base Library"
 PROJECT_NUMBER         = 1.0
 PROJECT_BRIEF          = 
-PROJECT_LOGO           = ../doc/img/redpitaya-logo.png
-OUTPUT_DIRECTORY       = ../doc/doxygen
+OUTPUT_DIRECTORY       = doxygen
 CREATE_SUBDIRS         = NO
 OUTPUT_LANGUAGE        = English
 BRIEF_MEMBER_DESC      = YES
@@ -107,7 +106,7 @@ WARN_LOGFILE           =
 #---------------------------------------------------------------------------
 # Configuration options related to the input files
 #---------------------------------------------------------------------------
-INPUT                  = rp.h
+INPUT                  = include/redpitaya/rp.h
 INPUT_ENCODING         = UTF-8
 FILE_PATTERNS          = *.c \
                          *.cc \
