+++ api/Makefile
@@ -9,7 +9,7 @@
 # REVISION ?= devbuild
 # VER:=$(shell cat info/info.json | grep version | sed -e 's/.*:\ *\"//' | sed -e 's/-.*//')
 
-LIBRP=../lib/librp.so
+LIBRP=lib/librp.so
 
 all: $(LIBRP)
