+++ api/src/analog_mixed_signals.h
