+++ api/src/oscilloscope.h
